DJ-60 DJ-61 Errors and Modals

+++ DigitalJournal/src/main/java/de/dhbw/softwareengineering/controller/RecoverPasswordController.java
@@ -14,9 +14,9 @@ import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
-
 import java.util.UUID;
 
 import static de.dhbw.softwareengineering.utilities.Constants.applicationContext;
@@ -30,7 +30,6 @@ public class RecoverPasswordController {
      *
      * @param email
      * @param model
-     *
      * @return
      */
     @RequestMapping(value = "/recoverpassword", method = RequestMethod.POST)
@@ -40,21 +39,25 @@ public class RecoverPasswordController {
         model.addAttribute(new LoginUser());
         model.addAttribute(new ContactRequest());
 
+        if(!Constants.emailPattern.matcher(email).matches()){
+            model.addAttribute("recover", "invalidEmail");
+            return "home";
+        }
+
         applicationContext.refresh();
 
-            UserDAO userDAO = applicationContext.getBean(UserDAO.class);
-            User user       = userDAO.getUserByEMail(email);
+        UserDAO userDAO = applicationContext.getBean(UserDAO.class);
+        User user = userDAO.getUserByEMail(email);
+
+        if (user != null) {
 
             PasswordRecoveryRequestDAO recoveryRequestDAO = applicationContext.getBean(PasswordRecoveryRequestDAO.class);
             // Delete all requests from user
-            for(PasswordRecoveryRequest recoveryRequest : recoveryRequestDAO.getAllRequestsFromUser(user.getUsername())){
+            for (PasswordRecoveryRequest recoveryRequest : recoveryRequestDAO.getAllRequestsFromUser(user.getUsername())) {
                 recoveryRequestDAO.deleteRequest(recoveryRequest);
                 Constants.prettyPrinter.info("Deleted password recovery request from user: " + recoveryRequest.getUsername());
             }
 
-        applicationContext.close();
-
-        if(user != null){
             model.addAttribute("recover", "true");
             model.addAttribute("email", email);
 
@@ -64,21 +67,19 @@ public class RecoverPasswordController {
             String[] recipients = {user.getEmail()};
 
 
-
             PasswordRecoveryRequest recoveryRequest = new PasswordRecoveryRequest();
-                                    recoveryRequest.setUsername(user.getUsername());
-                                    recoveryRequest.setRecoveryUUID(uuid);
-                                    recoveryRequest.setCreationDate(System.currentTimeMillis());
+            recoveryRequest.setUsername(user.getUsername());
+            recoveryRequest.setRecoveryUUID(uuid);
+            recoveryRequest.setCreationDate(System.currentTimeMillis());
 
             prettyPrinter.info(prettyPrinter.formatObject(recoveryRequest));
 
             recoveryRequestDAO.addRequest(recoveryRequest);
-
             Email.getInstance().sendEmailSSL(recipients, "DigitalJournal: Change your password", getEmailBody(url, user.getUsername()));
-        } else{
+        } else {
             model.addAttribute("recover", "false");
         }
-
+        applicationContext.close();
         return "home";
     }
 
@@ -87,23 +88,23 @@ public class RecoverPasswordController {
     public String index(@RequestParam(value = "uuid") String uuid, Model model, HttpSession session) {
 
         applicationContext.refresh();
-            PasswordRecoveryRequestDAO recoveryRequestDAO = applicationContext.getBean(PasswordRecoveryRequestDAO.class);
-            PasswordRecoveryRequest passwordRecoveryRequest = recoveryRequestDAO.getRequestByUUID(uuid);
+        PasswordRecoveryRequestDAO recoveryRequestDAO = applicationContext.getBean(PasswordRecoveryRequestDAO.class);
+        PasswordRecoveryRequest passwordRecoveryRequest = recoveryRequestDAO.getRequestByUUID(uuid);
 
-            prettyPrinter.info(prettyPrinter.formatObject(passwordRecoveryRequest));
+        prettyPrinter.info(prettyPrinter.formatObject(passwordRecoveryRequest));
 
-            User user = null;
-            if(passwordRecoveryRequest != null){
-                UserDAO userDAO = applicationContext.getBean(UserDAO.class);
-                user = userDAO.getUserByName(passwordRecoveryRequest.getUsername());
-            }
+        User user = null;
+        if (passwordRecoveryRequest != null) {
+            UserDAO userDAO = applicationContext.getBean(UserDAO.class);
+            user = userDAO.getUserByName(passwordRecoveryRequest.getUsername());
+        }
         applicationContext.close();
 
         if (passwordRecoveryRequest == null) {
             model.addAttribute("status", "notFound");
             return "error";
         } else {
-            if(session.getAttribute("changePasswordUser")!=null){
+            if (session.getAttribute("changePasswordUser") != null) {
                 session.removeAttribute("changePasswordUser");
             }
             session.setAttribute("changePasswordUser", user);
@@ -115,40 +116,44 @@ public class RecoverPasswordController {
     }
 
     @RequestMapping(value = "/changepassword", method = RequestMethod.POST)
-    public String index(@RequestParam("password") String password, @RequestParam("password_confirm") String password_confirm, ModelMap model, HttpSession session) {
+    public String index(@RequestParam("password") String password, @RequestParam("password_confirm") String password_confirm, ModelMap model, HttpSession session, RedirectAttributes redir) {
         // security checks
-        if(password.isEmpty() || password_confirm.isEmpty()){
-            model.addAttribute(Constants.STATUS_ATTRIBUTE_NAME, Constants.STATUSCODE_EMPTYFORM); return "changepassword";
-        } else if(!password.matches(password_confirm)){
-            model.addAttribute(Constants.STATUS_ATTRIBUTE_NAME, Constants.STATUSCODE_PWMISSMATCH); return "changepassword";
+        if (password.isEmpty() || password_confirm.isEmpty()) {
+            model.addAttribute(Constants.STATUS_ATTRIBUTE_NAME, Constants.STATUSCODE_EMPTYFORM);
+            return "changepassword";
+        } else if (!password.matches(password_confirm)) {
+            model.addAttribute(Constants.STATUS_ATTRIBUTE_NAME, Constants.STATUSCODE_PWMISSMATCH);
+            return "changepassword";
         } else if (password.length() < 6) {
-            model.addAttribute(Constants.STATUS_ATTRIBUTE_NAME, Constants.STATUSCODE_PWTOOSHORT); return "changepassword";
+            model.addAttribute(Constants.STATUS_ATTRIBUTE_NAME, Constants.STATUSCODE_PWTOOSHORT);
+            return "changepassword";
         } else if (password.length() > 42) {
-            model.addAttribute(Constants.STATUS_ATTRIBUTE_NAME, Constants.STATUSCODE_PWTOOLONG); return "changepassword";
+            model.addAttribute(Constants.STATUS_ATTRIBUTE_NAME, Constants.STATUSCODE_PWTOOLONG);
+            return "changepassword";
         }
 
         // get logged in user
         User user = (User) session.getAttribute(Constants.SESSION_CHANGEPWUSER);
 
         // change password of user and write to database
-        if(user != null){
+        if (user != null) {
             BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
             user.setPassword(encoder.encode(password));
 
             applicationContext.refresh();
-                UserDAO userDAO = applicationContext.getBean(UserDAO.class);
-                        userDAO.updateUser(user);
+            UserDAO userDAO = applicationContext.getBean(UserDAO.class);
+            userDAO.updateUser(user);
             applicationContext.close();
 
-            model.addAttribute(Constants.STATUS_ATTRIBUTE_NAME,Constants.STATUSCODE_PWCHANGESUCCESS);
-        }else{
+          // model.addAttribute("recover", Constants.STATUSCODE_PWCHANGESUCCESS);
+            redir.addFlashAttribute("recover", Constants.STATUSCODE_PWCHANGESUCCESS);
+        } else {
             return "error";
         }
 
-        if(session.getAttribute(Constants.SESSION_CHANGEPWUSER)!=null){
+        if (session.getAttribute(Constants.SESSION_CHANGEPWUSER) != null) {
             session.removeAttribute(Constants.SESSION_CHANGEPWUSER);
         }
-
         return "redirect:/";
     }
